Initial revision of Qualcomm Krait optimized cpuburn program
Primarily stresses the VABA instructions, which are dual-issued
with VLD2 instructions in the NEON pipeline. The ARM pipeline
is busy executing the SUBS instruction and the conditional
branch. The whole loop body consisting of 6 instructions gets
executed per 2 cycles, reaching the peak theoretical IPC=3.

Large loop unrolling is a bad idea on Krait because the
instructions decoder would suffer from cache line crossing
penalty and we would not be able to reach the peak IPC=3.
The loop body itself is small because this reduces the distance
between conditional branches. See
    https://plus.google.com/+SiarheiSiamashka/posts/CFnFYddbs7f
for more details about the Krait pipeline.

On the IFC6410 board clocked at 1.33GHz, the current draw from
a 5V PSU gets increased by 1150mA. For comparison, a multithreaded
litecoin miner (https://github.com/pooler/cpuminer) increases the
current draw by 850mA when run in the same conditions.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>